+++ lib/src/commonMain/kotlin/at/florianschuster/store/Store.kt
@@ -30,7 +30,6 @@ operator fun <Action> Store<*, Action, *>.component2(): (Action) -> Unit = ::dis
  * @param initialState The initial [State] of the [Store].
  * @param environment The [Environment] that is available within the contexts of [Reducer]s and [Effect]s.
  * @param effectScope The [CoroutineScope] in which [Effect]s are executed.
- * @param initialEffect An optional [Effect] that is executed when the [Store] is initialized.
  * @param reducer The [Reducer] that handles updating [State] based on dispatched [Action]s.
  * @param events An optional [StoreEvents] that emits [StoreEvent]s. Can be used for logging or testing.
  */
@@ -38,14 +37,12 @@ fun <Environment, Action, State> Store(
     initialState: State,
     environment: Environment,
     effectScope: CoroutineScope,
-    initialEffect: Effect<Environment, Action>? = null,
     reducer: Reducer<Environment, Action, State>,
     events: StoreEvents? = null,
 ): Store<Environment, Action, State> = StoreImplementation(
     effectScope = effectScope,
     environment = environment,
     initialState = initialState,
-    initialEffect = initialEffect,
     reducer = reducer,
     events = events,
 )
@@ -59,7 +56,6 @@ fun <Environment, Action, State> Store(
  * @param initialState The initial [State] of the created parent [Store].
  * @param environment The [Environment] that is available within the contexts of [Reducer]s and [Effect]s.
  * @param effectScope The [CoroutineScope] in which [Effect]s are executed.
- * @param initialEffect An optional [Effect] that is executed when the [Store] is initialized.
  * @param delegates A list of [DelegateStore]s that this [Store] delegates [Action]s and [State] to.
  * The order of the list of [delegates] is important, as it determines the order in which they receive
  * dispatched [Action]s.
@@ -69,14 +65,12 @@ fun <Environment, Action, State> Store(
     initialState: State,
     environment: Environment,
     effectScope: CoroutineScope,
-    initialEffect: Effect<Environment, Action>? = null,
     delegates: List<DelegateStore<Environment, Action, State, *, *, *>>,
     events: StoreEvents? = null,
 ): Store<Environment, Action, State> = DelegatingStoreImplementation(
     initialState = initialState,
     environment = environment,
     effectScope = effectScope,
-    initialEffect = initialEffect,
     delegates = delegates,
     events = events,
 )