+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # changelog
 
+## `[0.2.0]` - 2025-08-24
+
+- Remove `initialEffect` from `Store` factory functions
+- Add `initialEffect` to `Reducer` factory function
+
 ## `[0.1.0]` - 2025-08-03
 
 - initial release